# docs/openapi.yaml
...
openapi: 3.0.3
info:
  title: MCT Client Portal API
  description: Maine CyberTech Client Portal API. Bearer token authentication required unless noted as public.
  version: 1.0.0
  contact:
    name: MCT Support
    email: support@mainecybertech.com
servers:
  - url: https://api.mainecybertech.com
    description: Production
  - url: https://api.mainecybertech.us
    description: Dev
  - url: http://localhost:4000
    description: Local development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /webhooks/jira:
    post:
      tags:
        - Webhooks
      summary: Jira webhook handler
      responses:
        '200':
          description: Success
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
      security: []
  /webhooks/jsm:
    post:
      tags:
        - Webhooks
      summary: JSM webhook handler
      responses:
        '200':
          description: Success
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
      security: []
  /webhooks/m365:
    post:
      tags:
        - Webhooks
      summary: M365 webhook handler
      responses:
        '200':
          description: Success
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
      security: []
    get:
      tags:
        - Webhooks
      summary: M365 OAuth verification challenge
      responses:
        '200':
          description: Success
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
      security: []
  /webhooks/stripe:
    post:
      tags:
        - Webhooks
      summary: Stripe webhook handler
      responses:
        '200':
          description: Success
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
      security: []